[System Browser] Consolidate the test TCP listener with the product one
